Moved some stuff around
